add forge classic (#79)
* Update segsmaker.py

* Update setup.py

* z

* Delete LICENSE

* MIT

* f

* s

* k

* Update README.md

* d

* asdasd

+++ script/SM/venv.py
@@ -18,7 +18,8 @@ cwd = Path.cwd()
 
 vnv_FF = tmp / 'venv-fusion'
 vnv_SDT = tmp / 'venv-sd-trainer'
-vnv_C = tmp / 'venv-comfyui'
+vnv_C = tmp / 'venv-comfy-swarm'
+vnv_FC = tmp / 'python311'
 vnv_D = tmp / 'venv'
 
 SyS = get_ipython().system
@@ -29,42 +30,22 @@ def aDel():
         if name in globals():
             del globals()[name]
 
-def load_config():
-    config = json.load(MARK.open('r')) if MARK.exists() else {}
-    ui = config.get('ui')
+def trashing():
+    dirs1 = ['A1111', 'Forge', 'ReForge', 'Forge-Classic', 'ComfyUI', 'SwarmUI', 'FaceFusion', 'SDTrainer']
+    dirs2 = ['ckpt', 'lora', 'controlnet', 'svd', 'z123']
 
-    if ui == 'FaceFusion':
-        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/venv-fusion.tar.lz4'
-        need_space = 13 * 1024**3
-        vnv = vnv_FF
-    elif ui == 'SDTrainer':
-        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/venv-sd-trainer.tar.lz4'
-        need_space = 14 * 1024**3
-        vnv = vnv_SDT
-    elif ui == 'ComfyUI':
-        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/venv-comfyui.tar.lz4'
-        need_space = 14 * 1024**3
-        vnv = vnv_C
-    else:
-        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/venv-torch251-cu121-SSL.tar.lz4'
-        need_space = 14 * 1024**3
-        vnv = vnv_D
+    paths = [HOME / name for name in dirs1] + [tmp / name for name in dirs2]
+    for path in paths:
+        cmd = f'find {path} -type d -name .ipynb_checkpoints -exec rm -rf {{}} +'
+        SyS(f'{cmd} >/dev/null 2>&1')
 
-    fn = Path(url).name
-    return ui, url, need_space, vnv, fn
-
-def unused_venv():
-    if any(venv.exists() for venv in [vnv_FF, vnv_SDT, vnv_D]):
-        vnv_list = {
-            vnv_FF: [vnv_SDT, vnv_D, vnv_C],
-            vnv_SDT: [vnv_FF, vnv_D, vnv_C],
-            vnv_C: [vnv_FF, vnv_D, vnv_SDT],
-            vnv_D: [vnv_FF, vnv_SDT, vnv_C]
-        }.get(vnv)
-
-        if vnv_list:
-            rmf = f'rm -rf {" ".join(f"{venv}/* {venv}" for venv in vnv_list)}'
-            SyS(rmf)
+def check_pv():
+    try:
+        cmd = 'pv -V'
+        subprocess.run(shlex.split(cmd), capture_output=True, text=True, check=True)
+    except (subprocess.CalledProcessError, FileNotFoundError):
+        cmd = 'mamba install -y pv'
+        subprocess.run(shlex.split(cmd), stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
 def check_venv(folder):
     return int(du[0].split()[0]) if (du := get_ipython().getoutput(f'du -s -b {folder}')) else 0
@@ -89,32 +70,43 @@ def removing(directory, req_space):
         freed_space += size
     return freed_space
 
-def trashing():
-    dirs1 = ['A1111', 'Forge', 'ComfyUI', 'ReForge', 'FaceFusion', 'SDTrainer', 'SwarmUI']
-    dirs2 = ['ckpt', 'lora', 'controlnet', 'svd', 'z123']
+def unused_venv(vnv):
+    vl = [vnv_FF, vnv_SDT, vnv_D, vnv_C, vnv_FC]
+    vd = [v for v in vl if v != vnv and v.exists()]
+    if vd:
+        rmf = f"rm -rf {' '.join(f'{v}/* {v}' for v in vd)}"
+        SyS(rmf)
 
-    paths = [HOME / name for name in dirs1] + [tmp / name for name in dirs2]
-    for path in paths:
-        cmd = f'find {path} -type d -name .ipynb_checkpoints -exec rm -rf {{}} +'
-        SyS(f'{cmd} >/dev/null 2>&1')
+def load_config():
+    config = json.load(MARK.open('r')) if MARK.exists() else {}
+    ui = config.get('ui')
 
-def check_pv():
-    try:
-        cmd = 'pv -V'
-        subprocess.run(shlex.split(cmd), capture_output=True, text=True, check=True)
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        cmd = 'mamba install -y pv'
-        subprocess.run(shlex.split(cmd), stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+    if ui == 'FaceFusion':
+        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/venv-fusion.tar.lz4'
+        need_space = 13 * 1024**3
+        vnv = vnv_FF
+    elif ui == 'SDTrainer':
+        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/venv-sd-trainer.tar.lz4'
+        need_space = 14 * 1024**3
+        vnv = vnv_SDT
+    elif ui in ['ComfyUI', 'SwarmUI']:
+        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/SSL-ComfyUI-SwarmUI-Torch260-cu124.tar.lz4'
+        need_space = 14 * 1024**3
+        vnv = vnv_C
+    elif ui == 'Forge-Classic':
+        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/SSL-FC-Python311-Torch260-cu124.tar.lz4'
+        need_space = 14 * 1024**3
+        vnv = vnv_FC
+    else:
+        url = 'https://huggingface.co/gutris1/webui/resolve/main/env/SSL-Torch260-cu124.tar.lz4'
+        need_space = 14 * 1024**3
+        vnv = vnv_D
 
-def venv_exists(vnv, ui):
-    if vnv.exists():
-        size = check_venv(vnv)
-        if (ui == 'FaceFusion' and size < 7 * 1024**3) or size > 7 * 1024**3:
-            return True
-    return False
+    fn = Path(url).name
+    return ui, url, need_space, vnv, fn
 
-def install_venv(ui, url, need_space, fn):
-    unused_venv()
+def install_venv(ui, url, need_space, vnv, fn):
+    unused_venv(vnv)
     check_pv()
 
     clear_output(wait=True)
@@ -129,28 +121,31 @@ def install_venv(ui, url, need_space, fn):
             if req_space > 0: req_space -= removing(path, req_space)
 
     CD(tmp)
-    say('<b>【{red} Installing VENV{d} 】{red}</b>')
+    msg = 'Installing Forge Classic Python 3.11.13' if ui == 'Forge-Classic' else f'Installing {ui} VENV'
+    say(f'<b>【{{red}} {msg}{{d}} 】{{red}}</b>')
     download(url)
 
     SyS(f'pv {fn} | lz4 -d | tar xf -')
     Path(fn).unlink()
 
-    for cmd in [
-        f'rm -f {vnv}/bin/pip* {vnv}/bin/python*',
-        f'python3 -m venv {vnv}',
-        f'{pip} install -U --force-reinstall pip',
-        f'{pip} install ipykernel matplotlib',
-        f'{pip} uninstall -y ngrok pyngrok'
-    ]: SyS(f'{cmd} >/dev/null 2>&1')
+    if ui != 'Forge-Classic':
+        for cmd in [
+            f'rm -f {vnv}/bin/pip* {vnv}/bin/python*',
+            f'python3 -m venv {vnv}',
+            f'{pip} install -U --force-reinstall pip',
+            f'{pip} install ipykernel matplotlib',
+            f'{pip} uninstall -y ngrok pyngrok'
+        ]: SyS(f'{cmd} >/dev/null 2>&1')
 
-print('checking venv...')
+print('checking env...')
 tempe()
+
 ui, url, need_space, vnv, fn = load_config()
 pip = f'{vnv}/bin/python3 -m pip'
 
-if not venv_exists(vnv, ui): install_venv(ui, url, need_space, fn)
+if not vnv.exists(): install_venv(ui, url, need_space, vnv, fn)
 
+clear_output(wait=True)
 trashing()
 aDel()
-clear_output(wait=True)
 CD(cwd)
